backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
python-multipart==0.0.6
pillow==10.1.0
python-magic==0.4.27
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pytesseract==0.3.10
pyzbar==0.1.9
zxing-cpp==2.3.0
anthropic==0.7.8
openai==1.3.7
isbnlib==3.10.14
requests==2.31.0
cryptography==41.0.7
pytest==7.4.3
pytest-asyncio==0.21.1google-generativeai>=0.8.0
